# docs/swagger.yaml
definitions:
...
  ChapterResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/model.Chapter'
        type: array
    type: object
  CitiesResponse:
    properties:
      data:
        items:
          type: string
        type: array
    type: object
...
  CountriesResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/Country'
        type: array
    type: object
...
  CreatePositionInput:
    properties:
      code:
        type: string
      name:
        type: string
    required:
    - code
    - name
    type: object
  CreateStackInput:
    properties:
      avatar:
        type: string
      code:
        type: string
      name:
        type: string
    required:
    - code
    - name
    type: object
...
  ErrorResponse:
    properties:
      error:
        type: string
      errors:
        items:
          $ref: '#/definitions/view.ApiError'
        type: array
    type: object
...
  GetCurrenciesResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/Currency'
        type: array
    type: object
...
  GetQuestionResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/view.Question'
        type: array
    type: object
...
  MessageResponse:
    properties:
      message:
        type: string
    type: object
...
  OrganizationsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/model.Organization'
        type: array
    type: object
...
  PositionResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/model.Position'
        type: array
    type: object
...
  RolesResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/model.Role'
        type: array
    type: object
...
  SeniorityResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/model.Seniority'
        type: array
    type: object
...
  StackResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/model.Stack'
        type: array
    type: object
...
  UpdatePositionBody:
    properties:
      code:
        type: string
      name:
        type: string
    type: object
...
  model.Chapter:
    properties:
      code:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: string
      lead_id:
        type: string
      name:
        type: string
      updatedAt:
        type: string
    type: object
...
  view.ApiError:
    description: validation error details
    properties:
      enums:
        description: available options incase of field's payload is enums
        items:
          type: string
        type: array
      field:
        description: the field cause the error
        type: string
      msg:
        description: error message
        type: string
    type: object
...
info:
  contact:
    email: benjamin@d.foundation
    name: Nam Nguyen
    url: https://d.foundation
  description: This is api document for fortress project.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: FORTRESS API DOCUMENT
  version: v0.1.39
paths:
...
  /metadata/chapters:
    get:
      consumes:
      - application/json
      description: Get list values for chapters
      operationId: getChaptersList
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ChapterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get list values for chapters
      tags:
      - Metadata
  /metadata/countries:
    get:
      consumes:
      - application/json
      description: Get all countries
      operationId: getCountriesList
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CountriesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get all countries
      tags:
      - Metadata
  /metadata/countries/{country_id}/cities:
    get:
      consumes:
      - application/json
      description: Get list cities by country
      operationId: getCitiesList
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CitiesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get list cities by country
      tags:
      - Metadata
  /metadata/currencies:
    get:
      consumes:
      - application/json
      description: Get list currencies
      operationId: getCurrenciesList
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GetCurrenciesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get list currencies
      tags:
      - Metadata
  /metadata/organizations:
    get:
      consumes:
      - application/json
      description: Get list values for organizations
      operationId: getOrganizationsList
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/OrganizationsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get list values for organizations
      tags:
      - Metadata
  /metadata/positions:
    get:
      consumes:
      - application/json
      description: Get list values for positions
      operationId: getPositionsList
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/PositionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get list values for positions
      tags:
      - Metadata
    post:
      consumes:
      - application/json
      description: Create new position
      operationId: createPosition
      parameters:
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/CreatePositionInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create new position
      tags:
      - Metadata
  /metadata/positions/{id}:
    delete:
      consumes:
      - application/json
      description: Delete position by ID
      operationId: deletePosition
      parameters:
      - description: Position ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete position by ID
      tags:
      - Metadata
    put:
      consumes:
      - application/json
      description: Update position information by ID
      operationId: updatePosition
      parameters:
      - description: Position ID
        in: path
        name: id
        required: true
        type: string
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/UpdatePositionBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update position information by ID
      tags:
      - Metadata
  /metadata/project-statuses:
    get:
      consumes:
      - application/json
      description: Get list values for project statuses
      operationId: getProjectStatusesList
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get list values for project statuses
      tags:
      - Metadata
  /metadata/questions:
    get:
      consumes:
      - application/json
      description: Get list question by category and subcategory
      operationId: getQuestionsList
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GetQuestionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get list question by category and subcategory
      tags:
      - Metadata
  /metadata/roles:
    get:
      consumes:
      - application/json
      description: Get list roles
      operationId: getRolesList
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/RolesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get list roles
      tags:
      - Metadata
  /metadata/seniorities:
    get:
      consumes:
      - application/json
      description: Get list values for sentitorities
      operationId: getSenioritiesList
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/SeniorityResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get list values for sentitorities
      tags:
      - Metadata
  /metadata/stacks:
    get:
      consumes:
      - application/json
      description: Get list values for stacks
      operationId: getStacksList
      parameters:
      - description: Keyword
        in: query
        name: keyword
        type: string
      - description: Page
        in: query
        name: page
        type: string
      - description: Size
        in: query
        name: size
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/StackResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get list values for stacks
      tags:
      - Metadata
    post:
      consumes:
      - application/json
      description: Create new stack
      operationId: createStack
      parameters:
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/CreateStackInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create new stack
      tags:
      - Metadata
...
  /metadata/working-status:
    get:
      consumes:
      - application/json
      description: Get list values for working status
      operationId: getWorkingStatusList
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get list values for working status
      tags:
      - Metadata
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"